# Reduce performance overhead of ResourceLeakDetector