@normanmaurer 2017-09-16
@carl-mastrangelo PTAL... After I had a look at your change (https://github.com/netty/netty/commit/b32cd26a961c71cc5ad0e9f88fa8c15fbfd89930) I noticed we can do even better 👍