@normanmaurer 2017-09-17
Another number to share is that running `mvn -Pleak clean package` in the handler module before this patch took over 10 mins on my machine... With this patch it takes 2 minutes ;)