@normanmaurer 2017-09-18
Cherry-picked into 4.1 (3c8c7fc7e9) and 4.0 (4560197923).